// src/types/napcat.ts
export interface OB11BaseEvent {
  time: number;
  self_id: OneBotId;
  post_type: string;
}

export interface OB11Message extends OB11BaseEvent {
  post_type: 'message' | 'message_sent';
  message_type: 'private' | 'group';
  message_id: OneBotMessageId;
  user_id: OneBotId;
  message: OB11MessageSegment[];
  raw_message: string;
  sender: OB11Sender;
  group_id?: OneBotId;
}

export interface OB11Sender {
  user_id: OneBotId;
  nickname: string;
  card?: string;
  role?: 'owner' | 'admin' | 'member';
}

export interface OB11MessageSegment {
  type: string;
  data: Record<string, unknown>;
}